+++ crates/client/eth/src/sync.rs
@@ -3,31 +3,28 @@ use crate::l1_gas_price::{gas_price_worker, L1GasPrices};
 use crate::state_update::state_update_worker;
 use starknet_types_core::felt::Felt;
 
-// use crate::cli::SyncParams;
-use alloy::primitives::Address;
-use anyhow::Context;
-use dc_db::db_metrics::DbMetrics;
-use dc_db::{DatabaseService, DeoxysBackend};
-use dc_metrics::MetricsRegistry;
-// use dc_sync::fetch::fetchers::FetchConfig;
-// use dc_sync::metrics::block_metrics::BlockMetrics;
-// use dc_telemetry::TelemetryHandle;
-use dp_utils::service::Service;
+use dc_db::DeoxysBackend;
+
 use futures::lock::Mutex;
 use std::sync::Arc;
-use std::time::Duration;
-use tokio::task::JoinSet;
 
-pub async fn sync(
+pub async fn l1_sync_worker(
     backend: &DeoxysBackend,
     eth_client: &EthereumClient,
     chain_id: Felt,
     l1_gas_prices: Arc<Mutex<L1GasPrices>>,
+    gas_price_sync_disabled: bool,
 ) -> anyhow::Result<()> {
     let state_update_fut = async { state_update_worker(backend, eth_client, chain_id).await };
-    let gas_price_fut = async { gas_price_worker(eth_client, l1_gas_prices, true).await };
 
-    tokio::try_join!(state_update_fut, gas_price_fut)?;
+    if gas_price_sync_disabled {
+        // Only run the state update worker if gas price sync is disabled
+        state_update_fut.await?;
+    } else {
+        // Run both workers if gas price sync is enabled
+        let gas_price_fut = async { gas_price_worker(eth_client, l1_gas_prices, true).await };
+        tokio::try_join!(state_update_fut, gas_price_fut)?;
+    }
 
     Ok(())
 }