+++ crates/node/src/service/sync.rs
@@ -1,16 +1,12 @@
 use crate::cli::SyncParams;
-use alloy::primitives::Address;
 use anyhow::Context;
 use dc_db::db_metrics::DbMetrics;
 use dc_db::{DatabaseService, DeoxysBackend};
-use dc_eth::client::EthereumClient;
-use dc_eth::l1_gas_price::L1GasPrices;
 use dc_metrics::MetricsRegistry;
 use dc_sync::fetch::fetchers::FetchConfig;
 use dc_sync::metrics::block_metrics::BlockMetrics;
 use dc_telemetry::TelemetryHandle;
 use dp_utils::service::Service;
-use futures::lock::Mutex;
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::task::JoinSet;
@@ -35,27 +31,9 @@ impl SyncService {
         metrics_handle: MetricsRegistry,
         telemetry: TelemetryHandle,
     ) -> anyhow::Result<Self> {
-        // TODO: create l1 metrics here
         let block_metrics = BlockMetrics::register(&metrics_handle)?;
         let db_metrics = DbMetrics::register(&metrics_handle)?;
         let fetch_config = config.block_fetch_config();
-
-        // let l1_endpoint = if !config.sync_l1_disabled {
-        //     if let Some(l1_rpc_url) = &config.l1_endpoint {
-        //         Some(l1_rpc_url.clone())
-        //     } else {
-        //         return Err(anyhow::anyhow!(
-        //             "❗ No L1 endpoint provided. You must provide one in order to verify the synced state."
-        //         ));
-        //     }
-        // } else {
-        //     None
-        // };
-
-        // let core_address = Address::from_slice(config.network.l1_core_address().as_bytes());
-        // let eth_client = EthereumClient::new(l1_endpoint.unwrap(), core_address, metrics_handle)
-        //     .await
-        //     .context("Creating ethereum client")?;
 
         Ok(Self {
             db_backend: Arc::clone(db.backend()),