+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## Next release
 
+- feat: l1 sync service
+- feat: gas price worker for l1
 - test: tests added for eth client and event subscription
-
 - feat: experimental block production and mempool
 - refactor: L1BlockMetric is intialized inside the EthereumClient new function
 - refactor: BlockMetrics divided in L1BlockMetrics and BlockMetrics